requirements.txt
numpy==1.24.4
opencv_python==4.8.1.78
opencv_python_headless==4.9.0.80
Pillow==10.3.0
scipy==1.10.1
torch==1.12.1
torchvision==0.13.1
